chore: reorganize shadow and responsive-color
