+++ packages/prop-controllers/src/responsive-color/index.ts
@@ -1,0 +1,1 @@
+export * from './responsive-color'
